New two-part internal representation of colour
The driver for this change is mainly to support spot colours: they need some different
hanlding from others. It should also allow extension to spot mixes: at this stage
the mechanisms are undeveloped but the path is clear.

As the color model and actual setting are separate, I've broken into two parts.
At some future stage we may want to worry about foreground/background but that
is not done yet. Partly this depends on whether the colour is strictly internal.

Separating the setting data using commas or similar is in some was attractive but
as we always need spaces in the \special it's not really that helpful. The same
would apply to storing as a prop: attractive in a conceptual sense but not efficient
at point-of-use, and with no real gain.

At the moment, there is still no sense of setting or managing colours, just picking
up LaTeX2e ones. As such, the complexities of separations have not been touched at
all yet.