MAke l3color-extra loadable
Various internal changes to make the code clearer, including ensuring that
the 'return' value of parsing is not stored as \l__color_current_tl where
that is not appropriate.

This now can create colours directly (hence loading), though the parser is
untested.

One obvious question at this stage is whether in package mode data should
be exchanged with (x)color.