Make l3doc aware of "kernel" prefix
This allows us to skip the warning about foreign internals in the case
where it is "kernel" and the flag is set for the current .dtx being a
kernel module. We may wish to consider other 'special' cases: perhaps
"debug".

Note at present I've set the .dtx files in l3kernel to ignore "kernel"
prefixes, but not any other code.