Document image module interfaces
Slightly odd as the module is in l3trial but impacts on the ccore
l3drivers. Howerver, as none of the code can be used without loading
l3image we should be fine :)

One new interface is added, to provide a way for drivers to add to the
bounding box cache.