Initialise l3draw work
This is a stub at present, with the idea being to implement something
more-or-less the same as pgf (the base layer of TikZ). The key
difference between the pgf code and the plan here is that we have the
FPU to build on, which should provide much more even handling of e.g.
small numbers.

The plan is to stay interface-compatible with pgf if possible: it's well
thought-out but also that may allow 'drop in' TikZ compatiblity level.
(There are a few places some patching might be required.)

As pgf is a mature and complex system, the aim is to build up part by
part. In particular, some more tricky features will likely be missing
for some time. This is already evident in the driver-level code added to
l3kernel: at present, this doesn't have pgf's 'protocol' (buffering)
approach to paths.

The initial files follow the structure of pgfcore, though the plan at
present is to combine all of the path functions into a single (slightly
long) source.

Getting to a working set up will require work in various places: there
will be a number of commits to get anywhere near that. See also above
concerning the fact that some code will be simplified, particularly in a
first pass.