Implement fp_to_decimal and fp_to_scientific for tuples
The fp_to_tl representation is hard to parse, so it is useful to have
a more parseable one such as fp_to_scientific or fp_to_decimal whose
components can easily be converted for instance to dimensions.
The main drawback is that "\fp_to_decimal:n { ... } pt" is not
guaranteed anymore to be a dimension; I suspect however that in general
one knows when to expect a tuple or single number as a result.