First steps in box inclusion in drawings
In contrast to pgf, we want to do everything using driver-level specials
rather than raw PostScript/PDF. That means that the interfaces need to
be adjusted to avoid using translations at the driver level: these have
to be done in TeX.

There are various corrections to the driver code here.

At present, the SVG path may be completely out-of-line.

Note that some skews seem to mislead dvips/ps2pdf, resulting in entire
pages rotating. That does not seem to be due to 'leaking' rotations, and
is therefore likely down to some auto-detection somewhere.