Reset graphics state at start of a drawing
This is (likely) the most predictable approach for drawing-in-drawing
cases, such as when a drawing appears in a box which is itself in a
drawing.

An alternative approach would be to do the reset when a box is included.
Likely this needs some discussion.

See https://tex.stackexchange.com/questions/420645/ for background here.