Continuing to remove cross-module \__int_ usage
This covers many (though not all) of the \__int cases.

In the main, where there is a single use in a module, I've gone for
\int_eval:n even if there is a performance loss. Where multiple uses are
present, I've favoured keeping primitive access. Both of these choices
may need to be looked at: there are performance issues in some cases but
not in others, I suspect.

We may want to think about \number: often we've used it to allow
\exp_after:wN to an integer in one step.

A few more of these to do yet, some of which need rewrites of bits and
pieces.