xparse: Allow spaces before trailing optionals
This adds a new "!" modifier to the syntax which applies to optional
arguments. At present, it only does anything if the optional argument is
also trailing: others are left alone. The logic of that is something
like

    \DeclareDocumentCommand{\foo}{mom}{...}
    \foo{bar} [baz] {bop}

has to find a second mandatory, and grabbing "[" here as #3 would be
pretty strange.

What happens with multiple trailing optionals is not entirely clear.
I've gone for "!" applying on a per-argument basis, but that is only one
way of setting it up.