Make l3clist correctly unbrace items when braces are not needed
When adding items to a comma list, spaces were removed but not braces,
even for items that did not need them.  Now the treatment is consistent
by having a common (internal) function to test whether an item should
be braced or not.  For instance,

    \clist_set:Nn \l_tmpa_clist { {a} }
    \clist_if_in:NnTF \l_tmpa_clist { a } {true} {false}

gave false but now gives true.