Avoid combining chars in 8-bit case change test for lt
The trailing combining chars aren't set up for inputenc, which means
with TL'18 LaTeX2e they are pointing to an error message. That causes
havoc in the case changer: the older version has them just as chars.

Almost certainly these codepoints can't be used in an 8-bit document as
the fact they apply *after* the char is not supported by the engine. So
rather than try to pick up the error, simply adjust the test to stick to
legitimate input.