Cutting down overful boxes in source3
This deals with more-or-less all of the ones dues to overlong code
lines. There are various places that perhaps might be handled by
re-writing some docs. More tricky are the non-trivial number that come
at the end-of-definition lines: they are outside of something that can
be changed in the source. There are also a reasonable number that are (I
think) in the syntax environments.