Improve benchmarking quality; add tests
Rather than always repeating the code 4^n times for some n, repeat it a
"more continuous" number of times.  This avoids strange jumps in timing
with the earlier code when code passed from one 4^n to the next.

The tests check that benchmarking code does not give completely crazy
results.  Hopefully it's lax enough to pass on Travis.