Add "fake" prefix names (see #450)
We have many constants such as \c_underscore_str.  Of course they belong
to the "str" module, but if someone were to write an "underscore" package
it would not necessarily be crazy for them to have some parameter
\g_underscore_tl to customize the package.  To avoid confusion, reserve
these names.  I've not reserved two, three, ..., thirty because we are
deprecating these integer constants.  Sorry for the last commit, it was
a bit messed up, now fixed.