Add the factorial function fact() to l3fp (see #297)
The gamma function is much harder because it is defined at non-integer
arguments of course.  I've opted for a function fact() to be used
in expressions because none of the other advanced functions are
provided as \fp_...:n.  I've hesitated between the full word factorial
and the shorter name, to be more in line with things like cos or atan.
This can easily be changed.