Improve \char_<target>_case:N and add str version
This came up after a discussion with Ulrike. The old behaviour was
somewhat loose in terms of the catcode of the tokens produced. In this
commit, that is addressed by ensuring that the output tokens are always
re-generated to have the correct catcode. It also adds a str version,
which will be quite a bit faster and which gives the correct behaviour
in \str_<target>_case:n - that was not the case before.