Define 2e interfaces for filename stack and backend loading
This commit complements the commit (not committed yet, so no hash):
  "Incorporate l3 interfaces for filename stack and backend loading"
in the latex2e repository that incorporates these interfaces in
\@pushfilename, \@popfilename (both in ltclass.dtx), and \document (in
ltfiles.dtx).

This commit doesn't depend on its pair in the latex2e repository: expl3
should work the same way it did before this commit.  The latex2e commit,
however, requires this version to work properly.