Make \cctab_gset:Nn start from IniTeX codes
This means some adjustment to \cctab_const:Nn and
to the description of catcode tables.

This is the first step of this change. The 'early expl3'
changes mean there will need to be adjustments to how
document-level catcodes are set up. There are various
possible approaches: probably they will need to be
set explicitly, which will require reading load-unicode-data.
As that change is more complex, it will be in a second step.
We will also have to consider babel, UTF-8 and upTeX
issues.