Overhaul Greek case changer
This is now based on the ICU data at https://sites.google.com/site/icusite/design/case/greek-upper.
At present, ή to Ή is not included as this needs to
be handled separately from everything else.
As a result, those tests are currently commented out.
There is one test I also do not understand the result of ("νομικοῦ ἢ διεθνοῦς") in the ICU list: I will seek advice.