Change how special cases such as \l_tmpa_tl are detected (fixes #450)
Now we have a fixed list of words (alignment, tmpa, ...) that cannot
end up as modules for the purpose of the l3doc index.  When indexing
\c_alignment_token or \l_tmpa_seq or similar, the last word is taken
as the module instead of the first word.  But for \l_mymodule_tl,
since "mymodule" is not in our list, the variable ends up classified
as belonging to mymodule.