Refine documentation of \u in l3regex
Document that \u allows any variable, not just tl.  Remove the idea
of allowing \u as the number of repetitions of something (in the
sense that "a{2,\u{l_my_int}}" could denote repeating a from 2 to
\l_my_int times) because this can be done more cleanly by
x-expanding "a{2,\int_use:N\l_my_int}" before using it as a regex.