Suppress capturing groups in composed regex (see #590)
Previously, if the regex inserted with \ur included capturing groups,
then these were showing up as groups in the larger regex, thus shifing
all subsequent groups: for instance in "\ur{l_tmpa_regex}a(b?)" the
group number corresponding to "(b?)" would depend on the regex
\l_tmpa_regex.  Now all groups in the inserted regex are treated as
non-capturing groups and do not shift the group number.  We could
presumably add an option like \urg or \ur[capture] to keep capturing
the groups in the inserted regex.