implement faster group_align_safe_begin:
The code in this commit uses non-standard category codes and as a result
will not work correctly if group_align_safe_begin: is expanded once and
after that tokens with category code 2 get normalized. This should be
rarely the case, but in \unravel this would break for instance. As a
result this code might be considered unstable.