Properly correct xfrac issue
The problem here is tricky. The older set-up had the value of "numerator-top-sep" as part of the restricted template, so it was always set. However, one of the top or bottom sep values have to be maxdimen to avoid an error. With restrictions actually working, we couldn't undo the global setting of the top-sep, so we were stuck. Thus here I've made that a setting which is applied to an instance, not a restriction.