Move case data to two-part table system
This is step one of revision here. This commit loads the data and
adjusts the string and char functions, but does not
change the text case changing:
that will be in a later commit.

Renames and deprecations are also needed in
this area, and again will be handled in separate commits,