Fix \peek_analysis_map_inline:n for not-expanded tokens (fix #1113)
When followed by an implicit special token, the code saves the
next token using \futurelet, then stringifies the first token and
collects characters until hitting the saved next token.  If that
token was in a "notexpanded" state then the collection loop never
terminated.  Now fixed by simply running \futurelet twice to get
rid of the "notexpanded" state.  One could conceivably want to
keep track of it and restore it.