Improve o-expansion of the output of \peek_analysis_map_inline:n
Previously, we produced \exp_not:N \token for an expandable \token,
which meant that o-expanding that would give a version of \token
temporarily set to \relax.  Thus, \exp_after:wN \if_meaning:w #1
would not match as expected.  Now we do
\__kernel_exp_not:w \exp_after:wN { \exp_not:N #1 }
to avoid that issue yet still deal with the possibility of outer
macros.