Treat x-type as deprecated following normal policy
This means
- Retain all x-type variants present in last l3kernel
   release (2023-08-29)
- Remove variants from documentation
- Log the remove in l3obsolete

The \exp_args:N... family is excluded here as
they are needed internally in a few places:
a second pass will be done for these.

I have also ensured that in line with the
same policy position, all o-type variants present
in 2023-08-29 are retained: formally a separate thing
but goes together in terms of doing the work.