Correct \flag_clear:N error checking when debugging
I opted to only raise an expandable (rather than a more verbose
protected) error when the argument is undefined, because this allows
using the same auxiliary as for \flag_height:N and other expandable
flag commands.  This auxiliary is needed to clean up a bit and avoid
having infinitely many errors in a loop.