List a few more types (fparray, intarray, regex, token) in expl3.dtx
I'm not including here some things that are arguably types but are
not used as last parts of variables' names:

- Constant markers (quark, scan mark).
- Functions, mostly constant (cs, conditional, \prg_new_map_functions).
- Mostly-constant with nested names: msg, keys.
- Used in expressions: fp variables, fp functions, color, color_model.
- Other: pdf_object, flag.

The first three items here are most often constant so it makes sense to
not treat them as variables.  The next item consists of things like
colors or fp variables which are used next to "red" or "pi" that
certainly should be n-type.  It is not clear whether pdf_object and
flags should be turned to N-type variables.