Document, implement, and test doubly-linked implementation of prop
The hard part was surprisingly to implement \prop_show:N and log, which
have to tediously check every single detail of the internal structure of
the linked property list, in a way that should be fully robust to a
broken data structure.  The order in which functions are defined has
been modified, making the diff somewhat ugly, sorry.  The new test is
basically a concatenation of all the other tests, but with linked
property lists instead of flat ones.