Avoid overeager tokenization by \peek_analysis_map_inline:n (see #1434)
Now the peeking loop only ever looks at the next token and not the
one after.  This gives wrong results in rare cases (implicit catcode
1,2,10 char whose csname starts with another csname with the same
meaning) instead of only doing so in extra-rare cases (same with
extra conditions on the next-next token).  We prioritize having the
simplest effect on tokenization as some uses require that