Make peek_analysis_map_inline work for active spaces (fixes #1539)
At some point in the code the active character is taken as an argument
after being turned into a string, but a string space is simply skipped
when seeking macro arguments.  Now fixed by a trailing brace group
that gives an empty argument, and a trailing space that replaces the
argument in a followup \int_value:w `#1 ~ construction.